+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Map selections can now be exported as stitched PNG images - including the whole map!
 - Exported PNG images now properly utilize Sub, Up, Average and Paeth filters, resulting in better compression.
 - Added UV layer inspection on the model-viewer. Preview a texture from the ribbon and choose a UV layer to overlay.
+- Model textures can now be exported directly from the texture ribbon when previewing a model.
 
 ### Settings
 - Added option that controls the inclusion of prefix bones in glTF exports (thanks thunderysteak!).